recipe/auth/requirements.txt
PyJWT==1.6.1
PyYAML==3.12
